+++ helper/socket_error_other.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package main
+
+import "syscall"
+
+const peerResetError = syscall.ECONNRESET